% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arog_function.R
\name{sim}
\alias{sim}
\title{Simulating data from finite mixture of regression models}
\usage{
sim(x, prob, beta, ssd)
}
\arguments{
\item{x}{annotation matrix (no of SNPs * no of annotations)}

\item{prob}{component probabilities}

\item{beta}{regression coefficient  matrix ((1 + no of annotations)  * no of components)}

\item{ssd}{component standard deviations}
}
\value{
a response vector generated by finite mixture of regression modelling#'
}
\description{
This function allows you to simulate responses of finite mixture of linear regression models
}
\examples{
\dontrun{
n <- 100
p <- 50
x <- matrix(rnorm(n*p),n,p)
beta <- cbind(c(0,rep(3,5),rep(0,p-5)),c(0,rep(-1,5),rep(0,p-5)))
prob <- c(0.5,0.5)
ssd <- c(0.5,0.5)
y <- sim(x=x,ssd=ssd,beta=beta,prob=prob)
}
}
\author{
Sunyoung Shin\email{shin@stat.wisc.edu}
}
\keyword{arog}
\keyword{data}
\keyword{simulation}

